requirements.txt
celery
Django
django-rest-passwordreset
djangorestframework
drf-spectacular
model-bakery
psycopg2-binary
pytest
pytest-django
PyYAML
redis
requests
ujson
